# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentPay API
  version: 0.4.0
  description: |
    Seller commerce gateway, approval, transaction, and dispute API. AgentPay is a working name.
    Version 0.4.0 is the locked production target contract for M7.1. The current
    M7 Go runtime remains on the documented development compatibility baseline
    until the dependent LCH implementation tasks land. Production deployment
    must not advertise or partially enable this target contract before every
    authorization and persistence precondition is implemented.
    Seller project keys are proprietary bootstrap credentials accepted only by
    POST /v1/integration-access-tokens; this is not an OAuth token endpoint.
    Lean V1 project-key installations use the required AgentPay connector.
    Direct remote MCP OAuth clients are deferred. Ordinary MCP requests use
    short-lived ES256 access tokens. Browser
    purchase sessions and buyer-agent keys are distinct from seller credentials.
    Production MCP commercial mutations additionally require a one-time grant
    issued through the authenticated seller browser/BFF boundary; caller-
    asserted confirmation fields are development compatibility only.
  x-agentpay-contract-stage: target-state
  x-agentpay-runtime-baseline: m7-development-preview
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{apiHost}
    description: Environment-specific API Gateway origin
    variables:
      apiHost:
...
paths:
...
  /v1/sellers:
    post:
      summary: Create a seller storefront
      tags: [Sellers]
      operationId: createSeller
      security: [{ sellerBearer: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSellerRequest' }
      responses:
        '201': { $ref: '#/components/responses/SellerCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
  /v1/sellers/{sellerId}/routes:
    get:
      summary: List paid routes owned by a seller
      tags: [Sellers]
      operationId: listPaidRoutes
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
      responses:
        '200':
          description: Seller-owned paid routes
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
    post:
      summary: Add a paid route to a seller
      tags: [Sellers]
      operationId: createPaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRouteRequest' }
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
  /v1/sellers/{sellerId}/routes/{routeId}:
    get:
      summary: Get one seller-owned paid route
      tags: [Sellers]
      operationId: getPaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
      responses:
        '200':
          description: Seller-owned paid route
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
    patch:
      summary: Update the price used by future purchase intents
      tags: [Sellers]
      operationId: updatePaidRoutePrice
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoutePriceRequest' }
      responses:
        '200':
          description: Route price updated; existing unexpired purchase intents keep their frozen quote
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/sellers/{sellerId}/routes/{routeId}/validation:
    get:
      summary: Validate one paid route against current publication requirements
      tags: [Sellers]
      operationId: validatePaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
      responses:
        '200':
          description: Current deterministic validation result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteValidationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/sellers/{sellerId}/routes/{routeId}/publish:
    post:
      summary: Publish or resume a validated paid route
      tags: [Sellers]
      operationId: publishPaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteVersionRequest' }
      responses:
        '200':
          description: Published paid route
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/sellers/{sellerId}/routes/{routeId}/pause:
    post:
      summary: Pause a published paid route
      tags: [Sellers]
      operationId: pausePaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteVersionRequest' }
      responses:
        '200':
          description: Paused paid route
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/sellers/{sellerId}/routes/{routeId}/archive:
    post:
      summary: Permanently archive a non-published paid route
      tags: [Sellers]
      operationId: archivePaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteVersionRequest' }
      responses:
        '200':
          description: Archived paid route
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /v1/sellers/{sellerId}/routes/{routeId}/emergency-disable:
    post:
      summary: Immediately disable a published paid route
      tags: [Sellers]
      operationId: emergencyDisablePaidRoute
      security: [{ sellerBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/SellerId' }
        - { $ref: '#/components/parameters/RouteId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteVersionRequest' }
      responses:
        '200':
          description: Emergency-disabled paid route
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaidRoute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/PermissionDenied' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
  /store/{slug}/manifest.json:
    get:
      summary: Get a storefront manifest
      tags: [Storefront]
      operationId: getStorefrontManifest
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9-]+$' }
      responses:
        '200':
          description: Short-lived signed active storefront discovery document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedStorefrontManifest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Known storefront is inactive; the signed tombstone contains no products
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedStorefrontTombstone' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /store/{slug}/llms.txt:
    get:
      summary: Get the storefront agent-discovery document
      tags: [Storefront]
      operationId: getStorefrontLlmsText
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9-]+$' }
      responses:
        '200':
          description: Plain-text agent discovery document
          content:
            text/plain:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/SellerInactive' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
components:
  securitySchemes:
    sellerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Seller-scoped service credential derived from the authenticated web session by the trusted web/BFF boundary; it is never stored in browser JavaScript or accepted as a substitute for ownership checks.
    projectKey:
      type: apiKey
      in: header
      name: X-AgentPay-Project-Key
      description: Proprietary seller project bootstrap credential accepted only by the integration access-token exchange.
    buyerAgentKey:
      type: apiKey
      in: header
      name: X-AgentPay-Agent-Key
      description: Buyer-agent credential; never a seller project key.
    browserPurchaseCookie:
      type: apiKey
      in: cookie
      name: __Host-agentpay_purchase
      description: Opaque server-side browser grant. Commerce authority expires separately from longer-lived read and remediation access.
    approvalCookie:
      type: apiKey
      in: cookie
      name: __Host-agentpay_approvals
      description: Opaque browser grant set created by exchanging one or more approval invitation fragment tokens.
    stripeWebhookSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe-generated signature verified against the exact raw request body and the environment-specific webhook endpoint secret.
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' } }
    SellerId: { name: sellerId, in: path, required: true, schema: { type: string, pattern: '^sel_[A-Za-z0-9]+$' } }
    RouteId: { name: routeId, in: path, required: true, schema: { type: string, pattern: '^rte_[A-Za-z0-9]+$' } }
...
  schemas:
...
    CreateSellerRequest:
      type: object
      additionalProperties: false
      required: [name, slug, upstreamBaseUrl]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        slug: { type: string, pattern: '^[a-z0-9-]{3,48}$' }
        upstreamBaseUrl: { type: string, format: uri }
...
    CreateRouteRequest:
      type: object
      additionalProperties: false
      required: [displayName, productSlug, method, pathPattern, description, mimeType, amount, asset, network, payTo]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
          description: Seller-approved human product name; surrounding and repeated whitespace is normalized.
        productSlug:
          type: string
          minLength: 3
          maxLength: 80
          pattern: '^[A-Za-z0-9](?:[A-Za-z0-9 _-]*[A-Za-z0-9])?$'
          description: Seller-approved public slug input normalized to lowercase single-hyphen form and unique within the seller.
        method: { type: string, enum: [GET, POST] }
        pathPattern: { type: string, pattern: '^/[A-Za-z0-9/_-]+$' }
        description: { type: string, minLength: 1, maxLength: 500 }
        mimeType: { type: string, maxLength: 120 }
        amount: { $ref: '#/components/schemas/PositiveAtomicAmount' }
        asset: { type: string, minLength: 1, maxLength: 160 }
        network: { type: string, minLength: 1, maxLength: 80 }
        payTo: { type: string, minLength: 1, maxLength: 160 }
        approvalThresholdAmount: { oneOf: [{ $ref: '#/components/schemas/AtomicAmount' }, { type: 'null' }] }
        upstreamTimeoutSeconds: { type: integer, minimum: 1, maximum: 30, default: 20 }
        publishImmediately:
          type: boolean
          default: true
          description: Publication is attempted only after the draft is persisted; all current publication checks run and failure leaves the route unpublished.
...
    PaidRoute:
      allOf:
        - { $ref: '#/components/schemas/CreateRouteRequest' }
        - type: object
          required: [routeId, sellerId, lifecycleStatus, enabled, createdAt, updatedAt, version]
          properties:
            routeId: { type: string }
            sellerId: { type: string }
            displayName: { type: string, minLength: 1, maxLength: 120 }
            productSlug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', minLength: 3, maxLength: 80 }
            lifecycleStatus: { type: string, enum: [draft, published, paused, archived, emergency_disabled] }
            enabled: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            version: { type: integer, minimum: 1 }
    UpdateRoutePriceRequest:
      type: object
      additionalProperties: false
      required: [amount, expectedVersion]
      properties:
        amount: { $ref: '#/components/schemas/PositiveAtomicAmount' }
        expectedVersion: { type: integer, minimum: 1 }
    RouteVersionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
    RouteValidationResult:
      type: object
      additionalProperties: false
      required: [sellerId, routeId, valid, checks, version]
      properties:
        sellerId: { type: string }
        routeId: { type: string }
        valid: { type: boolean }
        checks:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [name, passed, message]
            properties:
              name: { type: string }
              passed: { type: boolean }
              message: { type: string }
        version: { type: integer, minimum: 1 }
...
    SignedStorefrontManifest:
      type: object
      additionalProperties: false
      required: [document, signature]
      properties:
        document: { $ref: '#/components/schemas/StorefrontManifest' }
        signature: { $ref: '#/components/schemas/DiscoverySignature' }
    SignedStorefrontTombstone:
      type: object
      additionalProperties: false
      required: [document, signature]
      properties:
        document: { $ref: '#/components/schemas/StorefrontTombstone' }
        signature: { $ref: '#/components/schemas/DiscoverySignature' }
...
    AtomicAmount: { type: string, pattern: '^[0-9]+$', example: '10000' }
    PositiveAtomicAmount: { type: string, pattern: '^[1-9][0-9]*$', example: '35000000' }
...
  responses:
    SellerCreated:
      description: Seller created
      content: { application/json: { schema: { $ref: '#/components/schemas/Seller' } } }
...
    BadRequest: { description: 400 invalid_request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized:
      description: 401 invalid_credential, token_expired, or token_revoked
      headers:
        WWW-Authenticate: { schema: { type: string } }
        Cache-Control: { schema: { type: string, const: no-store } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound: { description: '404 not_found, including concealed cross-tenant resources', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: '409 state_conflict, idempotency_conflict, payment_replayed, or token_replayed', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    SellerInactive: { description: 410 seller_inactive; known seller cannot participate in discovery or new commerce, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: 422 validation_failed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    PermissionDenied: { description: '403 subscription_inactive, insufficient_scope, or permission_denied', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited:
      description: Caller, credential, browser session, seller quota, or public endpoint rate limit is exhausted
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    DependencyUnavailable:
      description: 503 dependency_unavailable; authoritative authorization or transaction state cannot be established
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }